Move code out of LLMonFHIR to SpeziFHIR
