--- LLMonFHIR/FHIR Display/FHIRResourcesInstructionsView.swift
@@ -1,65 +1,0 @@
-//
-// This source file is part of the Stanford LLM on FHIR project
-//
-// SPDX-FileCopyrightText: 2023 Stanford University
-//
-// SPDX-License-Identifier: MIT
-//
-
-import SpeziFHIR
-import SwiftUI
-
-
-struct FHIRResourcesInstructionsView: View {
-    @Environment(FHIRStore.self) private var fhirStore
-    @AppStorage(StorageKeys.onboardingInstructions) private var onboardingInstructions = true
-    
-    
-    var body: some View {
-        if fhirStore.allResources.isEmpty {
-            Section {
-                VStack(alignment: .center) {
-                    Image(systemName: "doc.text.magnifyingglass")
-                        .accessibilityHidden(true)
-                        .font(.system(size: 90))
-                        .foregroundColor(.accentColor)
-                        .padding(.vertical, 8)
-                    Text("FHIR_RESOURCES_VIEW_NO_RESOURCES")
-                        .frame(maxWidth: .infinity)
-                        .multilineTextAlignment(.leading)
-                }
-            }
-        } else if onboardingInstructions {
-            Section {
-                VStack(alignment: .center) {
-                    HStack {
-                        Spacer()
-                        Button(
-                            action: {
-                                withAnimation {
-                                    onboardingInstructions = false
-                                }
-                            },
-                            label: {
-                                Image(systemName: "xmark")
-                                    .accessibilityLabel(Text("DISMISS_ONBOARDING_HINT"))
-                                    .foregroundColor(.secondary)
-                            }
-                        )
-                    }
-                        .padding(.horizontal, -8)
-                    Image(systemName: "hand.wave.fill")
-                        .accessibilityHidden(true)
-                        .font(.system(size: 75))
-                        .foregroundColor(.accentColor)
-                        .padding(.bottom, 8)
-                    Text("FHIR_RESOURCES_VIEW_INSTRUCTION")
-                        .frame(maxWidth: .infinity)
-                        .multilineTextAlignment(.leading)
-                }
-            }
-        } else {
-            EmptyView()
-        }
-    }
-}
